Rename file section.vue to Section.vue
