+++ pages/legal-notices.vue
@@ -29,7 +29,7 @@
 
 <script>
 import { documents, documentFetcher } from '~/services/document-fetcher'
-import SectionSlice from '~/components/slices/section'
+import SectionSlice from '@/components/slices/Section'
 export default {
   name: 'LegalNotices',
   nuxtI18n: {
